Include other processes in time measure
time.time() gives time from system wall-clock so
it also counts the time for other processes.

time.clock() is good for benchmarking, because it
measures only time of our process, but we are not benchmarking.

What we want is simply terminate before time limit in competition
environment. Thus, looking at system wall-clock is safer.