Remove adding random epsilon to UCT
Random epsilon was added to break ties in a random manner.

The benefit of this is that UCT in the very beginning doesn't 
start always with the same moves but with random.
On the long run other starting moves will also be sampled,
thus ties will be broken anyway.

The downsides of adding random epsilon are less performance
and more nondeterministic behaviour (for testing, debugging this 
is a downside).

reference implementation on mcts.ai doesn't add random epsilon.