Better heuristic
Problem with before was that the center field was worth 4 points,
the same as 4 (!) fields near the border, so it was focusing too much
on center.

Now every field is worth 101, 102, 103, or 104 (the center),
which leads tiny bit towards center, but not at the cost of empty
fields, which are usually crucial.