Remove Python implementation
Python implementation is about 2 orders of magnitude slower than C++.
Because of that, in typical competition or testing setting (where for example 1
second for move is given), C++ will be always chosen over Python.
So, it's better to focus only on C++ implementation and do not spend
time and duplicate everything in Python implementation.