Remove player symbols from Algorithm
It was redundant and confusing, now code is simpler.

Only State knows what is the current and next player.
Now State has to implement get_enemy() instead of it 
being global, unelegant, static one.

Algorithms get a reference to state, so it can always ask
"who is the current player" or "give me the opponent of X".