Pass copy of current_state in Tester
Algorithms can modify given state, it can be rubbish in the end
of processing, their job is to just return move.

It's convienent in Algorithms to just modify received state
by make_move etc. assuming the caller will create a copy (if needed).

Otherwise, every Algorithm would have to create a copy,
which sometimes it's not needed, plus it will be repeated in every
Algorithm.

To not lose current state, throw-away copies should be passed.