Fix MCTS implementation, no UCT for now
Problem was that select() was always maximizng, for both players,
which is wrong, the second player should be minimazing.

Another problem was that in simulation it returned 1 even if opponent
won the simulation.

Since asymetric expansion with UCT formula is a bit more complicated,
but the bug was in the foundation of MCTS, to simplify things, this
commits drops UCT. It will be added in the future. But now the most
important is to get even pure MCTS right.

So now I believe it's correct, it selects child to simulate randomly
and increases the counters, propagating them up to the root. 

In the end it takes the move to the node in first level with the 
best win ratio.