Improve MCTS implementation, no children
No children table for each state (node).
Now states have only a pointer to the parent, no children at all.

Before all the children were allocated, which was a waste of time
and memory, since most of them stayed not visited.

Now, in alignment with pseudocode from
"Monte-Carlo Tree Search and Rapid Action Value
Estimation in Computer Go", maximum one child per simulation
can be allocated and added to the tree.

To check in O(1) if a node is already in the tree, hash table is used.