Fix MCTS, states are the same if parent is the same
A state can be reached from 2 (or more) parents.
Such state was not added to the tree (was taken from hashmap),
which was incorrect, because in propagation wrong path was updated.

Now, == and hash() was updated, so that two states are the same
only if they have the same board AND the same parent.

Also added more TicTacToe tests.

MAX_SIMULATIONS is now global (for tests and for library) and is 10k.

DRAW_SCORE was changed from 0.5 to 0.9 for TicTacToe.

Also improved a lot quality of pseudo random number generation,
Mersenne Twister + uniform distribution, not rand() with modulo.