Move transposition_table to Minimax
Like tree_table is in MCTS, the same transposition_table
is specific for Minimax, so it goes to Minimax algorithm,
not generic State. It can also be allocated on stack and
doesn't have to be static, simpler.

Also drop inserting hard-coded moves into TT, it wasn't 
too helpful and hash codes of State changed (because
hash() changed). If somebody needs to insert stuff to TT,
it's easy to manually edit the code to do that.