Children inside every State instead one tree_table
It's about 25% faster this way and more "traditional", common,
easier to understand. 

When getting child, now they're directly in the
state (pointers to them), so it's easy and fast. 

Before one had to take a parent state,
make a move (maybe with a copy before...), take it from global
hashmap and undo a move.