Remove move from State
Was redundant and not updated (in undo_move). 
This also makes things a bit simpler (no need to worry about it in 
make_move and undo_move). Also making State a bit lighter.

Nowhere in the code we need to know *only* from State, what was
the move leading to that State. We always have some State and we
are considering some moves (so we already know them), no need to 
store them within a State (at least with current code).