Add block flag to MCTS
If it's a game in which blocking is possible, one should pass block = true.
By default: false. Blocking is possible in TicTacToe and Connect 4, which
means that the winning move of an opponent is the best move for us.

Isola doesn't have this property though. Move that is winning for an
enemy is not a valid move for us. So we shouldn't waste time for searching
blocking moves.