Add max_moves to Minimax
Maximum best max_moves will be taken into consideration in a Minimax.

Assuming moves are sensibly sorted from best to worse this reduces
the branching factor, which allows to search deeper and improves quality
of play.

Tested on Isola, with max_moves = 72, it won 64/100 games against not
limited version. Assuming binomial distribution this gives p-value of 0.006,
which means 99% confidence that this is statististically significant improvement.