Print cutBF statistic
The average number of successors considered at a node where a cut-off
occurs. Only nodes where at least one move is searched (i.e., not
just a transposition) are counted.

It gives a rough feeling of how many nodes we have to expand to have
a cut-off. The closer to 1, the better.