Log instead of cout in get_move()
So that one can print logs it or save them to file, at convenient time.

stdout is usually used to print out the correct move on competitions
sites, not the debug logs, so redirecting them instantly there is not the
best idea.

Also removed verbose flag passed to algorithms as now it's not needed.
The logs are always saved and it's up to programmer how he will use them
after calling get_move().