Restore order of moves
This goes to depth 7 on first move instead of 6.

Reasoning why this is better move ordering:

Step_moves are ordered quite well, centeric ones are first.
So, we can take "the best" step move and try all the remove moves for it.

Remove moves are sorted in a bit worse order (there are many more 
sensible remove moves, so it's harder to order them as good as step moves).

Instead taking "probably not the best" remove move and trying all step moves.