Count only moves around the player
Before it was doing it in the distance of 2 around the player.
This turned out to be almost the same as looking only in the
distance of 1:

Match 2038: 2 Minimax won
1 Minimax wins: 1038
2 Minimax wins: 1000
Draws: 0
Ratio: 0.509323
Lower confidence bound: 0.480559
Upper confidence bound: 0.538043

Minimax 1 looked only at available moves around the player.

Thus, because of code simplification it's better to look only 
at depth 1.