Ability to play against any executable
To play with standalone bots, Executable algorithm was added.

It gets move by running a given command, for example ./wiman.o
(launching a process of external bot which waits for the input), 
passing state input to the bot and reading its move.

What's nice about it is that the external bot can be written in any
language. It only has to read state in external_format and output
its move (in a format understood by Move's read function).